main: confirming delete guest

+++ app/src/main/java/com/example/convidados/view/viewholder/GuestViewHolder.kt
@@ -1,6 +1,8 @@
 package com.example.convidados.view.viewholder
 
+import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.RecyclerView
+import com.example.convidados.R
 import com.example.convidados.databinding.RowGuestBinding
 import com.example.convidados.model.GuestModel
 import com.example.convidados.view.listener.OnGuestListener
@@ -16,7 +18,15 @@ class GuestViewHolder(private val bind: RowGuestBinding, private val listener: O
         }
 
         bind.textName.setOnLongClickListener {
-            listener.onDelete(guest.id)
+            AlertDialog.Builder(itemView.context)
+                .setTitle(R.string.delete_guest_title)
+                .setMessage(R.string.delete_guest_message)
+                .setPositiveButton(R.string.yes) { dialog, wich ->
+                    listener.onDelete(guest.id)
+                }
+                .setNegativeButton(R.string.no, null)
+                .create()
+                .show()
             true
         }
     }
